backend/requirements.txt
torch~=1.11.0
Pillow~=9.1.0
pandas~=1.3.4
torchvision~=0.1.6
fastapi~=0.78.0
uvicorn~=0.17.6
numpy~=1.21.4
opencv-python~=4.5.4.58
scipy~=1.7.2
scikit-image~=0.19.2
python-multipart